+++ common/src/main/java/io/github/flemmli97/runecraftory/common/entities/BaseMonster.java
@@ -643,6 +643,11 @@ public abstract class BaseMonster extends PathfinderMob implements Enemy, IAnima
     }
 
     @Override
+    public boolean removeWhenFarAway(double distanceToClosestPlayer) {
+        return !this.isTamed();
+    }
+
+    @Override
     public ItemStack getPickResult() {
         return SpawnEgg.fromType(this.getType()).map(ItemStack::new).orElse(ItemStack.EMPTY);
     }