+++ common/src/main/java/io/github/flemmli97/runecraftory/common/world/structure/AmbrosiaForestStructure.java
@@ -29,7 +29,7 @@ public class AmbrosiaForestStructure extends StructureFeature<JigsawConfiguratio
 
         BlockPos blockpos = context.chunkPos().getMiddleBlockPosition(0);
         int topLandY = context.chunkGenerator().getFirstFreeHeight(blockpos.getX(), blockpos.getZ(), Heightmap.Types.WORLD_SURFACE_WG, context.heightAccessor());
-        blockpos = blockpos.above(topLandY - 2);
+        blockpos = blockpos.above(topLandY - 1);
         return JigsawPlacement.addPieces(
                 context,
                 PoolElementStructurePiece::new,