+++ fabric/src/main/java/io/github/flemmli97/runecraftory/fabric/config/MobConfigSpec.java
@@ -25,9 +25,12 @@ public class MobConfigSpec {
     public final CommentedJsonConfig.DoubleVal gateHealthGain;
     public final CommentedJsonConfig.DoubleVal gateDefGain;
     public final CommentedJsonConfig.DoubleVal gateMDefGain;
-    public final CommentedJsonConfig.IntVal spawnChance;
     public final CommentedJsonConfig.IntVal gateXP;
     public final CommentedJsonConfig.IntVal gateMoney;
+    public final CommentedJsonConfig.IntVal spawnChance;
+    public final CommentedJsonConfig.DoubleVal minDist;
+    public final CommentedJsonConfig.IntVal maxGroup;
+    public final CommentedJsonConfig.IntVal maxNearby;
 
     public final Map<ResourceLocation, EntityPropertySpecs> mobSpecs = new HashMap<>();
 
@@ -43,9 +46,13 @@ public class MobConfigSpec {
         this.gateHealthGain = builder.comment("Health gain per level of gates").defineInRange("Health Gain", 25, 0, Double.MAX_VALUE);
         this.gateDefGain = builder.comment("Defence gain per level of gates").defineInRange("Defence Gain", 5, 0, Double.MAX_VALUE);
         this.gateMDefGain = builder.comment("Magic defence gain per level of gates").defineInRange("Magic Defence Gain", 5, 0, Double.MAX_VALUE);
-        this.spawnChance = builder.comment("Chance for next spawn (1/x chance per tick)").defineInRange("Spawn", 150, 0, Integer.MAX_VALUE);
         this.gateXP = builder.comment("Base xp a gate gives").defineInRange("XP", 12, 0, Integer.MAX_VALUE);
         this.gateMoney = builder.comment("Money a gate gives").defineInRange("Money", 4, 0, Integer.MAX_VALUE);
+        this.spawnChance = builder.comment("Chance for next spawn (1/x chance per tick)").defineInRange("Spawn", 150, 0, Integer.MAX_VALUE);
+        this.minDist = builder.comment("Radius to check for other gates. If more than Max Group gates are in that radius no other gates will spawn").defineInRange("Min Dist", 48d, 0, Double.MAX_VALUE);
+        this.maxGroup = builder.comment("Max amount of gates in Min Dist radius that can exist").defineInRange("Max Group", 2, 0, Integer.MAX_VALUE);
+        this.maxNearby = builder.comment("If more than x mobs are near a gate the gate will not spawn more").defineInRange("Max Nearby", 4, 0, Integer.MAX_VALUE);
+
         builder.pop();
         for (Map.Entry<ResourceLocation, EntityProperties> e : MobConfig.propertiesMap.entrySet()) {
             builder.push(e.getKey().toString());