+++ forge/src/main/java/io/github/flemmli97/runecraftory/forge/config/MobConfigSpec.java
@@ -21,9 +21,12 @@ public class MobConfigSpec {
     public final ForgeConfigSpec.DoubleValue gateHealthGain;
     public final ForgeConfigSpec.DoubleValue gateDefGain;
     public final ForgeConfigSpec.DoubleValue gateMDefGain;
-    public final ForgeConfigSpec.IntValue spawnChance;
     public final ForgeConfigSpec.IntValue gateXP;
     public final ForgeConfigSpec.IntValue gateMoney;
+    public final ForgeConfigSpec.IntValue spawnChance;
+    public final ForgeConfigSpec.DoubleValue minDist;
+    public final ForgeConfigSpec.IntValue maxGroup;
+    public final ForgeConfigSpec.IntValue maxNearby;
 
     public final Map<ResourceLocation, EntityPropertySpecs> mobSpecs = new HashMap<>();
 
@@ -39,9 +42,12 @@ public class MobConfigSpec {
         this.gateHealthGain = builder.comment("Health gain per level of gates").defineInRange("Health Gain", 25, 0, Double.MAX_VALUE);
         this.gateDefGain = builder.comment("Defence gain per level of gates").defineInRange("Defence Gain", 5, 0, Double.MAX_VALUE);
         this.gateMDefGain = builder.comment("Magic defence gain per level of gates").defineInRange("Magic Defence Gain", 5, 0, Double.MAX_VALUE);
-        this.spawnChance = builder.comment("Chance for next spawn (1/x chance per tick)").defineInRange("Spawn", 150, 0, Integer.MAX_VALUE);
         this.gateXP = builder.comment("Base xp a gate gives").defineInRange("XP", 12, 0, Integer.MAX_VALUE);
         this.gateMoney = builder.comment("Money a gate gives").defineInRange("Money", 4, 0, Integer.MAX_VALUE);
+        this.spawnChance = builder.comment("Chance for next spawn (1/x chance per tick)").defineInRange("Spawn", 150, 0, Integer.MAX_VALUE);
+        this.minDist = builder.comment("Radius to check for other gates. If more than Max Group gates are in that radius no other gates will spawn").defineInRange("Min Dist", 48d, 0, Double.MAX_VALUE);
+        this.maxGroup = builder.comment("Max amount of gates in Min Dist radius that can exist").defineInRange("Max Group", 2, 0, Integer.MAX_VALUE);
+        this.maxNearby = builder.comment("If more than x mobs are near a gate the gate will not spawn more").defineInRange("Max Nearby", 4, 0, Integer.MAX_VALUE);
         builder.pop();
         for (Map.Entry<ResourceLocation, EntityProperties> e : MobConfig.propertiesMap.entrySet()) {
             builder.push(e.getKey().toString());